[refactor] 코드 리뷰 1차 반영
패키지 이름 변경
외부 시스템 예외 전송 모듈 예외 로깅 추가
