+++ src/main/java/com/prgrms/rg/web/common/GlobalControllerAdvice.java
@@ -1,11 +1,11 @@
-package com.prgrms.rg.web.global;
+package com.prgrms.rg.web.common;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.prgrms.rg.web.global.message.ExceptionMessageSender;
-import com.prgrms.rg.web.global.results.InternalServerErrorResult;
+import com.prgrms.rg.web.common.message.ExceptionMessageSender;
+import com.prgrms.rg.web.common.results.InternalServerErrorResult;
 
 @RestControllerAdvice
 public class GlobalControllerAdvice {
