+++ src/main/java/com/prgrms/rg/web/common/message/ExceptionMessageSender.java
@@ -1,4 +1,4 @@
-package com.prgrms.rg.web.global.message;
+package com.prgrms.rg.web.common.message;
 
 public interface ExceptionMessageSender {
 	void send(Exception exception);
